src/llm/fastapi/src/vectorstore/requirements.txt
langchain==0.2.1
langchain-community==0.2.1
langserve==0.2.1
openai==1.30.5
fastapi==0.111.0
langchain-anthropic==0.1.15
psycopg2-binary==2.9.9
pgvector==0.2.5
tiktoken==0.7.0
boto3==1.34.117
llama-index==0.10.42
docx2txt==0.8
chardet